+++ tests/findAll.test.js
@@ -19,7 +19,6 @@
 const tap = require('tap')
 const abstractLogger = require('abstract-logging')
 const { MongoClient } = require('mongodb')
-const { omit: lomit } = require('lodash')
 
 const { STATES } = require('../lib/consts')
 const CrudService = require('../lib/CrudService')
@@ -32,12 +31,10 @@ const {
   getMongoDatabaseName,
   getMongoURL,
   BOOKS_COLLECTION_NAME,
-  RAW_PROJECTION_PLAIN_INCLUSIVE,
-  RAW_PROJECTION_PLAIN_EXCLUSIVE,
 } = require('./utils')
 
 const EMPTY_QUERY = {}
-const ONLY_ID_PROJECTION = []
+const DEFAULT_PROJECTION = { _id: 1 }
 const ALL_FIELDS = Object.keys(publicFixtures[0])
 
 const context = {
@@ -87,7 +84,7 @@ tap.test('findAll', async t => {
     t.test('could return an empty array', async t => {
       t.plan(1)
 
-      const data = await crudService.findAll(context, { foo: 'bar' }, ONLY_ID_PROJECTION).toArray()
+      const data = await crudService.findAll(context, { foo: 'bar' }, DEFAULT_PROJECTION).toArray()
 
       t.strictSame(data, [])
     })
@@ -107,7 +104,7 @@ tap.test('findAll', async t => {
     t.test('could return an empty array', async t => {
       t.plan(1)
 
-      const data = await crudService.findAll(context, { foo: 'bar' }, ONLY_ID_PROJECTION).toArray()
+      const data = await crudService.findAll(context, { foo: 'bar' }, DEFAULT_PROJECTION).toArray()
 
       t.strictSame(data, [])
     })
@@ -134,7 +131,7 @@ tap.test('findAll', async t => {
     t.test('should return only _id if no projection is specified', async t => {
       t.plan(1)
 
-      const data = await crudService.findAll(context, EMPTY_QUERY, ONLY_ID_PROJECTION).toArray()
+      const data = await crudService.findAll(context, EMPTY_QUERY, DEFAULT_PROJECTION).toArray()
 
       t.strictSame(data, publicFixtures.map(mapOnlyId))
     })
@@ -142,7 +139,7 @@ tap.test('findAll', async t => {
     t.test('should return only _id without projection', async t => {
       t.plan(1)
 
-      const data = await crudService.findAll(context, EMPTY_QUERY).toArray()
+      const data = await crudService.findAll(context, EMPTY_QUERY, DEFAULT_PROJECTION).toArray()
 
       t.strictSame(data, publicFixtures.map(mapOnlyId))
     })
@@ -154,7 +151,7 @@ tap.test('findAll', async t => {
     t.test('should return with the correct order ascendant', async t => {
       t.plan(1)
 
-      const data = await crudService.findAll(context, EMPTY_QUERY, ONLY_ID_PROJECTION, { price: 1 }).toArray()
+      const data = await crudService.findAll(context, EMPTY_QUERY, DEFAULT_PROJECTION, { price: 1 }).toArray()
       t.strictSame(data, publicFixtures
         .sort((a, b) => sortByPrice(a, b, 1))
         .map(mapOnlyId)
@@ -164,7 +161,7 @@ tap.test('findAll', async t => {
     t.test('should return with the correct order descendent', async t => {
       t.plan(1)
 
-      const data = await crudService.findAll(context, EMPTY_QUERY, ONLY_ID_PROJECTION, { price: -1 }).toArray()
+      const data = await crudService.findAll(context, EMPTY_QUERY, DEFAULT_PROJECTION, { price: -1 }).toArray()
 
       t.strictSame(data, publicFixtures
         .sort((a, b) => sortByPrice(a, b, -1))
@@ -175,7 +172,7 @@ tap.test('findAll', async t => {
     t.test('should return with the correct order ascendant for multiple fields', async t => {
       t.plan(1)
 
-      const data = await crudService.findAll(context, EMPTY_QUERY, ONLY_ID_PROJECTION, { price: 1, name: 1 }).toArray()
+      const data = await crudService.findAll(context, EMPTY_QUERY, DEFAULT_PROJECTION, { price: 1, name: 1 }).toArray()
       t.strictSame(data, publicFixtures
         .sort((a, b) => sortByName(a, b, 1))
         .sort((a, b) => sortByPrice(a, b, 1))
@@ -189,7 +186,7 @@ tap.test('findAll', async t => {
       const data = await crudService.findAll(
         context,
         EMPTY_QUERY,
-        ONLY_ID_PROJECTION,
+        DEFAULT_PROJECTION,
         { price: -1, name: -1 }
       ).toArray()
       t.strictSame(data, publicFixtures
@@ -202,7 +199,7 @@ tap.test('findAll', async t => {
     t.test('should return with the correct order: ascendant for the first field, descendent for the second', async t => {
       t.plan(1)
 
-      const data = await crudService.findAll(context, EMPTY_QUERY, ONLY_ID_PROJECTION, { price: 1, name: -1 }).toArray()
+      const data = await crudService.findAll(context, EMPTY_QUERY, DEFAULT_PROJECTION, { price: 1, name: -1 }).toArray()
       t.strictSame(data, publicFixtures
         .sort((a, b) => sortByName(a, b, -1))
         .sort((a, b) => sortByPrice(a, b, 1))
@@ -213,7 +210,7 @@ tap.test('findAll', async t => {
     t.test('should return with the correct order: descendent for the first field, ascendant for the second', async t => {
       t.plan(1)
 
-      const data = await crudService.findAll(context, EMPTY_QUERY, ONLY_ID_PROJECTION, { price: -1, name: 1 }).toArray()
+      const data = await crudService.findAll(context, EMPTY_QUERY, DEFAULT_PROJECTION, { price: -1, name: 1 }).toArray()
       t.strictSame(data, publicFixtures
         .sort((a, b) => sortByName(a, b, 1))
         .sort((a, b) => sortByPrice(a, b, -1))
@@ -227,7 +224,7 @@ tap.test('findAll', async t => {
       const data = await crudService.findAll(
         context,
         EMPTY_QUERY,
-        ONLY_ID_PROJECTION,
+        DEFAULT_PROJECTION,
         { price: -1, name: 1, publishDate: 1 }
       ).toArray()
       t.strictSame(data, publicFixtures
@@ -244,7 +241,7 @@ tap.test('findAll', async t => {
       const data = await crudService.findAll(
         context,
         EMPTY_QUERY,
-        ONLY_ID_PROJECTION,
+        DEFAULT_PROJECTION,
         { price: -1, 'attachments.name': 1, publishDate: 1 }
       ).toArray()
       t.strictSame(data, publicFixtures
@@ -267,7 +264,7 @@ tap.test('findAll', async t => {
       const data = await crudService.findAll(
         context,
         EMPTY_QUERY,
-        ONLY_ID_PROJECTION,
+        DEFAULT_PROJECTION,
         { price: 1 },
         SKIP_COUNT
       ).toArray()
@@ -291,7 +288,7 @@ tap.test('findAll', async t => {
       const data = await crudService.findAll(
         context,
         EMPTY_QUERY,
-        ONLY_ID_PROJECTION,
+        DEFAULT_PROJECTION,
         { price: 1 },
         0,
         LIMIT
@@ -315,7 +312,7 @@ tap.test('findAll', async t => {
       const data = await crudService.findAll(
         context,
         EMPTY_QUERY,
-        ONLY_ID_PROJECTION,
+        DEFAULT_PROJECTION,
         { price: 1 },
         0,
         500,
@@ -340,7 +337,7 @@ tap.test('findAll', async t => {
             },
           },
         },
-        ONLY_ID_PROJECTION,
+        DEFAULT_PROJECTION,
         { price: 1 },
         0,
         500,
@@ -354,98 +351,6 @@ tap.test('findAll', async t => {
           })
       }
       t.strictSame(data, fixtures.filter(mf).map(mapOnlyId))
-    })
-  })
-
-  t.test('rawProjection', t => {
-    t.plan(4)
-
-    const SORT = undefined
-    const SKIP = 0
-    const LIMIT = 20
-
-    const expectedDocs = fixtures.map((doc) => {
-      return doc.attachments
-        ? {
-          _id: doc._id,
-          isbn: doc.isbn,
-          price: doc.price,
-          attachments: doc.attachments,
-        }
-        : {
-          _id: doc._id,
-          isbn: doc.isbn,
-          price: doc.price,
-        }
-    })
-
-    const expectedDocsWithExcludedFields = fixtures.map((doc) => {
-      const docCopy = { ...doc }
-      return lomit(docCopy, ['attachments', 'isbn', 'price'])
-    })
-
-    t.test('should return only the right fields', async t => {
-      t.plan(1)
-
-      const data = await crudService.findAll(
-        context,
-        EMPTY_QUERY,
-        [...ONLY_ID_PROJECTION, RAW_PROJECTION_PLAIN_INCLUSIVE],
-        SORT,
-        SKIP,
-        LIMIT,
-        [STATES.DRAFT, STATES.PUBLIC, STATES.TRASH, STATES.DELETED]
-      ).toArray()
-
-      t.strictSame(data, expectedDocs)
-    })
-
-    t.test('should override fields in _p parameter', async t => {
-      t.plan(1)
-
-      const data = await crudService.findAll(
-        context,
-        EMPTY_QUERY,
-        ['attachments', RAW_PROJECTION_PLAIN_INCLUSIVE],
-        SORT,
-        SKIP,
-        LIMIT,
-        [STATES.DRAFT, STATES.PUBLIC, STATES.TRASH, STATES.DELETED]
-      ).toArray()
-
-      t.strictSame(data, expectedDocs)
-    })
-
-    t.test('should exclude specified fields', async t => {
-      t.plan(1)
-
-      const data = await crudService.findAll(
-        context,
-        EMPTY_QUERY,
-        [RAW_PROJECTION_PLAIN_EXCLUSIVE],
-        SORT,
-        SKIP,
-        LIMIT,
-        [STATES.DRAFT, STATES.PUBLIC, STATES.TRASH, STATES.DELETED]
-      ).toArray()
-
-      t.strictSame(data, expectedDocsWithExcludedFields)
-    })
-
-    t.test('should exclude specified fields overridding existing _p specification', async t => {
-      t.plan(1)
-
-      const data = await crudService.findAll(
-        context,
-        EMPTY_QUERY,
-        ['attachments', RAW_PROJECTION_PLAIN_EXCLUSIVE],
-        SORT,
-        SKIP,
-        LIMIT,
-        [STATES.DRAFT, STATES.PUBLIC, STATES.TRASH, STATES.DELETED]
-      ).toArray()
-
-      t.strictSame(data, expectedDocsWithExcludedFields)
     })
   })
 })