+++ tests/patchById.test.js
@@ -31,6 +31,7 @@ const {
   getMongoDatabaseName,
   getMongoURL,
   BOOKS_COLLECTION_NAME,
+  getProjectionFromObject,
 } = require('./utils')
 
 const {
@@ -43,6 +44,7 @@ const context = {
   userId: newUpdaterId,
   now: new Date('2018-02-08'),
 }
+const DEFAULT_PROJECTION = { _id: 1 }
 
 tap.test('patchById', async t => {
   const databaseName = getMongoDatabaseName()
@@ -62,7 +64,7 @@ tap.test('patchById', async t => {
   // eslint-disable-next-line prefer-destructuring
   const chosenDoc = fixtures[1]
   const chosenDocId = chosenDoc._id
-  const chosenDocProjection = Object.keys(chosenDoc)
+  const chosenDocProjection = getProjectionFromObject(chosenDoc)
   const updateCommand = () => ({ $set: { price: 44.0 } })
   const updatedDoc = {
     ...fixtures[1],
@@ -138,7 +140,7 @@ tap.test('patchById', async t => {
 
     const matchingQuery = { price: { $lt: 0 } }
 
-    const r = await crudService.patchById(context, fixtures[4]._id, updateCommand(), matchingQuery)
+    const r = await crudService.patchById(context, fixtures[4]._id, updateCommand(), matchingQuery, DEFAULT_PROJECTION)
 
     t.test('should return null', t => {
       t.plan(1)
@@ -160,7 +162,7 @@ tap.test('patchById', async t => {
       fixtures[4]._id,
       updateCommand(),
       matchingQuery,
-      {},
+      DEFAULT_PROJECTION,
       [STATES.DRAFT]
     )
 
@@ -247,7 +249,7 @@ tap.test('patchById', async t => {
     veryBadCommands.forEach((testConf) => {
       t.test(JSON.stringify(testConf.cmd), async t => {
         try {
-          await crudService.patchById(context, chosenDocId, testConf.cmd)
+          await crudService.patchById(context, chosenDocId, testConf.cmd, {}, DEFAULT_PROJECTION)
           t.fail()
         } catch (error) {
           t.ok(testConf.regex.test(error.message), error.message)
@@ -283,7 +285,7 @@ tap.test('patchById', async t => {
   const arrayQuery = { 'attachments.name': 'note' }
   const [doc] = fixtures
   const docId = doc._id
-  const docProjection = Object.keys(doc)
+  const docProjection = getProjectionFromObject(doc)
 
   t.test('update nested object in array', async t => {
     t.plan(2)